+++ packages/button/src/button/Button.tsx
@@ -5,12 +5,10 @@ import type {
   BorderRadiusType,
   FontsSizesType,
   ShadowsType,
-  PikasUIContextType,
 } from '@pikas-ui/styles'
-import { PikasUIContext } from '@pikas-ui/styles'
+import { useTheme } from '@pikas-ui/styles'
 import { styled } from '@pikas-ui/styles'
 import type { AnchorHTMLAttributes, ButtonHTMLAttributes } from 'react'
-import { useContext } from 'react'
 import React, { forwardRef, useCallback } from 'react'
 import type { IconProps, IconStyleType } from '@pikas-ui/icons'
 import { BeatLoader } from '@pikas-ui/loader'
@@ -223,7 +221,6 @@ const getContent = ({
   contentColorHex,
   textTransform,
   gap,
-  pikasUiContext,
 }: {
   LeftIcon?: React.FC<IconProps>
   RightIcon?: React.FC<IconProps>
@@ -235,14 +232,15 @@ const getContent = ({
   contentColorHex?: string
   textTransform?: ButtonTextTransformType
   gap?: ButtonGapType
-  pikasUiContext: PikasUIContextType
 }): React.ReactNode => {
-  if (!pikasUiContext) return null
+  const theme = useTheme()
+
+  if (!theme) return null
   return (
     <>
       <LoadingContainer>
         <BeatLoader
-          size={pikasUiContext.fontSizes['EM-XX-SMALL'].value}
+          size={theme.fontSizes['EM-XX-SMALL'].value}
           colorHex={getContentColor({
             outlined,
             contentColorHex: contentColorHex,
@@ -317,7 +315,7 @@ export const Button = forwardRef<HTMLButtonElement, ButtonProps>(
     },
     ref
   ) => {
-    const pikasUiContext = useContext(PikasUIContext)
+    const theme = useTheme()
 
     const handleClick = useCallback((): void => {
       if (disabled || loading) {
@@ -327,7 +325,7 @@ export const Button = forwardRef<HTMLButtonElement, ButtonProps>(
       onClick?.()
     }, [disabled, onClick, loading])
 
-    if (!pikasUiContext) return null
+    if (!theme) return null
     return (
       <ButtonDOM
         ref={ref}
@@ -346,10 +344,10 @@ export const Button = forwardRef<HTMLButtonElement, ButtonProps>(
 
           ...getColors({
             outlined,
-            colorHex: colorHex || (color && pikasUiContext.colors[color].value),
+            colorHex: colorHex || (color && theme.colors[color].value),
             contentColorHex:
               contentColorHex ||
-              (contentColor && pikasUiContext.colors[contentColor].value),
+              (contentColor && theme.colors[contentColor].value),
           }),
 
           ...styles?.button,
@@ -357,7 +355,6 @@ export const Button = forwardRef<HTMLButtonElement, ButtonProps>(
         {...props}
       >
         {getContent({
-          pikasUiContext,
           LeftIcon,
           RightIcon,
           children,
@@ -366,10 +363,10 @@ export const Button = forwardRef<HTMLButtonElement, ButtonProps>(
           outlined,
           styles,
           textTransform,
-          colorHex: colorHex || (color && pikasUiContext.colors[color].value),
+          colorHex: colorHex || (color && theme.colors[color].value),
           contentColorHex:
             contentColorHex ||
-            (contentColor && pikasUiContext.colors[contentColor].value),
+            (contentColor && theme.colors[contentColor].value),
         })}
       </ButtonDOM>
     )
@@ -425,7 +422,7 @@ export const ButtonLink = forwardRef<HTMLAnchorElement, ButtonLinkProps>(
     },
     ref
   ) => {
-    const pikasUiContext = useContext(PikasUIContext)
+    const theme = useTheme()
 
     const handleClick = useCallback((): void => {
       if (disabled || loading) {
@@ -435,7 +432,7 @@ export const ButtonLink = forwardRef<HTMLAnchorElement, ButtonLinkProps>(
       onClick?.()
     }, [disabled, onClick, loading])
 
-    if (!pikasUiContext) return null
+    if (!theme) return null
     return (
       <ButtonDOM
         as="a"
@@ -455,17 +452,16 @@ export const ButtonLink = forwardRef<HTMLAnchorElement, ButtonLinkProps>(
 
           ...getColors({
             outlined,
-            colorHex: colorHex || (color && pikasUiContext.colors[color].value),
+            colorHex: colorHex || (color && theme.colors[color].value),
             contentColorHex:
               contentColorHex ||
-              (contentColor && pikasUiContext.colors[contentColor].value),
+              (contentColor && theme.colors[contentColor].value),
           }),
           ...styles?.button,
         }}
         {...props}
       >
         {getContent({
-          pikasUiContext,
           LeftIcon,
           RightIcon,
           children,
@@ -474,10 +470,10 @@ export const ButtonLink = forwardRef<HTMLAnchorElement, ButtonLinkProps>(
           outlined,
           styles,
           textTransform,
-          colorHex: colorHex || (color && pikasUiContext.colors[color].value),
+          colorHex: colorHex || (color && theme.colors[color].value),
           contentColorHex:
             contentColorHex ||
-            (contentColor && pikasUiContext.colors[contentColor].value),
+            (contentColor && theme.colors[contentColor].value),
         })}
       </ButtonDOM>
     )