+++ packages/button/src/buttonIcon/ButtonIcon.tsx
@@ -5,10 +5,8 @@ import type {
   ShadowsType,
   SizesType,
 } from '@pikas-ui/styles'
-import { PikasUIContext } from '@pikas-ui/styles'
-import { styled, Sizes } from '@pikas-ui/styles'
+import { styled, Sizes, useTheme } from '@pikas-ui/styles'
 import type { AnchorHTMLAttributes, ButtonHTMLAttributes } from 'react'
-import { useContext } from 'react'
 import React, { forwardRef, useCallback } from 'react'
 import type {
   ButtonTypeType,
@@ -247,7 +245,7 @@ export const ButtonIcon = forwardRef<HTMLButtonElement, ButtonIconProps>(
     },
     ref
   ) => {
-    const pikasUiContext = useContext(PikasUIContext)
+    const theme = useTheme()
 
     const handleClick = useCallback((): void => {
       if (disabled || loading) {
@@ -257,7 +255,7 @@ export const ButtonIcon = forwardRef<HTMLButtonElement, ButtonIconProps>(
       onClick?.()
     }, [disabled, onClick, loading])
 
-    if (!pikasUiContext) return null
+    if (!theme) return null
     return (
       <ButtonIconDOM
         ref={ref}
@@ -271,10 +269,10 @@ export const ButtonIcon = forwardRef<HTMLButtonElement, ButtonIconProps>(
 
           ...getColors({
             outlined,
-            colorHex: colorHex || (color && pikasUiContext.colors[color].value),
+            colorHex: colorHex || (color && theme.colors[color].value),
             contentColorHex:
               contentColorHex ||
-              (contentColor && pikasUiContext.colors[contentColor].value),
+              (contentColor && theme.colors[contentColor].value),
           }),
 
           ...styles?.button,
@@ -282,10 +280,10 @@ export const ButtonIcon = forwardRef<HTMLButtonElement, ButtonIconProps>(
         {...props}
       >
         {getContent({
-          colorHex: colorHex || (color && pikasUiContext.colors[color].value),
+          colorHex: colorHex || (color && theme.colors[color].value),
           contentColorHex:
             contentColorHex ||
-            (contentColor && pikasUiContext.colors[contentColor].value),
+            (contentColor && theme.colors[contentColor].value),
           loading,
           outlined,
           size,
@@ -335,7 +333,7 @@ export const ButtonIconLink = forwardRef<
     },
     ref
   ) => {
-    const pikasUiContext = useContext(PikasUIContext)
+    const theme = useTheme()
 
     const handleClick = useCallback((): void => {
       if (disabled || loading) {
@@ -345,7 +343,7 @@ export const ButtonIconLink = forwardRef<
       onClick?.()
     }, [disabled, onClick, loading])
 
-    if (!pikasUiContext) return null
+    if (!theme) return null
     return (
       <ButtonIconDOM
         as="a"
@@ -360,10 +358,10 @@ export const ButtonIconLink = forwardRef<
 
           ...getColors({
             outlined,
-            colorHex: colorHex || (color && pikasUiContext.colors[color].value),
+            colorHex: colorHex || (color && theme.colors[color].value),
             contentColorHex:
               contentColorHex ||
-              (contentColor && pikasUiContext.colors[contentColor].value),
+              (contentColor && theme.colors[contentColor].value),
           }),
 
           ...styles?.button,
@@ -371,10 +369,10 @@ export const ButtonIconLink = forwardRef<
         {...props}
       >
         {getContent({
-          colorHex: colorHex || (color && pikasUiContext.colors[color].value),
+          colorHex: colorHex || (color && theme.colors[color].value),
           contentColorHex:
             contentColorHex ||
-            (contentColor && pikasUiContext.colors[contentColor].value),
+            (contentColor && theme.colors[contentColor].value),
           loading,
           outlined,
           size,