+++ packages/textfield/src/Textfield.tsx
@@ -7,13 +7,11 @@ import type {
   FontsSizesType,
   BorderRadiusType,
 } from '@pikas-ui/styles'
-import { PikasUIContext } from '@pikas-ui/styles'
-import { styled } from '@pikas-ui/styles'
+import { styled, useTheme } from '@pikas-ui/styles'
 import { Description, Label, TextError } from '@pikas-ui/text'
 import * as LabelPrimitive from '@radix-ui/react-label'
 import fontColorContrast from 'font-color-contrast'
 import type { InputHTMLAttributes } from 'react'
-import { useContext } from 'react'
 import { forwardRef } from 'react'
 import React, { useRef, useState } from 'react'
 import useMergedRef from '@react-hook/merged-ref'
@@ -296,7 +294,7 @@ export const Textfield = forwardRef<HTMLInputElement, TextfieldProps>(
     const refInput = useRef<HTMLInputElement>(null)
     const multiRef = useMergedRef(ref, refInput)
     const [focus, setFocus] = useState(false)
-    const pikasUiContext = useContext(PikasUIContext)
+    const theme = useTheme()
 
     const onChangeInput = (e: React.ChangeEvent<HTMLInputElement>): void => {
       if (type === 'number' && refInput.current) {
@@ -324,9 +322,9 @@ export const Textfield = forwardRef<HTMLInputElement, TextfieldProps>(
       return (
         colorHex ||
         (color ? `$${color}` : undefined) ||
-        (pikasUiContext
+        (theme
           ? fontColorContrast(
-              pikasUiContext.colors[backgroundColor || 'WHITE'].value,
+              theme.colors[backgroundColor || 'WHITE'].value,
               0.7
             )
           : undefined) ||