+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @pikas-ui/docs
 
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @pikas-ui/button@2.1.2
+  - @pikas-ui/checkbox@2.1.2
+  - @pikas-ui/styles@2.3.2
+  - @pikas-ui/table@2.1.3
+  - @pikas-ui/textarea@2.1.3
+  - @pikas-ui/textfield@2.1.3
+  - @pikas-ui/tooltip@2.1.2
+  - @pikas-ui/searchbar@2.1.3
+  - @pikas-ui/alert@2.1.2
+  - @pikas-ui/avatar@1.0.2
+  - @pikas-ui/context-menu@2.1.3
+  - @pikas-ui/dropdown-menu@2.1.3
+  - @pikas-ui/icons@2.1.2
+  - @pikas-ui/loader@2.1.2
+  - @pikas-ui/select@3.1.3
+  - @pikas-ui/separator@2.1.2
+  - @pikas-ui/slider@2.1.2
+  - @pikas-ui/switch@1.1.2
+  - @pikas-ui/text@3.1.2
+
 ## 1.0.11
 
 ### Patch Changes