+++ packages/alert/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pikas-ui/alert
 
+## 2.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @pikas-ui/styles@2.3.2
+  - @pikas-ui/icons@2.1.2
+
 ## 2.1.1
 
 ### Patch Changes