+++ packages/avatar/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pikas-ui/avatar
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @pikas-ui/styles@2.3.2
+
 ## 1.0.1
 
 ### Patch Changes