+++ packages/button/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pikas-ui/button
 
+## 2.1.2
+
+### Patch Changes
+
+- Add useTheme hook
+- Updated dependencies
+  - @pikas-ui/styles@2.3.2
+  - @pikas-ui/icons@2.1.2
+  - @pikas-ui/loader@2.1.2
+
 ## 2.1.1
 
 ### Patch Changes