+++ packages/loader/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pikas-ui/loader
 
+## 2.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @pikas-ui/styles@2.3.2
+
 ## 2.1.1
 
 ### Patch Changes