+++ packages/switch/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pikas-ui/switch
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @pikas-ui/styles@2.3.2
+  - @pikas-ui/tooltip@2.1.2
+  - @pikas-ui/icons@2.1.2
+  - @pikas-ui/text@3.1.2
+
 ## 1.1.1
 
 ### Patch Changes