+++ packages/table/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pikas-ui/table
 
+## 2.1.3
+
+### Patch Changes
+
+- Add useTheme hook
+- Updated dependencies
+  - @pikas-ui/checkbox@2.1.2
+  - @pikas-ui/styles@2.3.2
+  - @pikas-ui/icons@2.1.2
+  - @pikas-ui/loader@2.1.2
+  - @pikas-ui/select@3.1.3
+
 ## 2.1.2
 
 ### Patch Changes