+++ packages/textarea/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pikas-ui/textarea
 
+## 2.1.3
+
+### Patch Changes
+
+- Add useTheme hook
+- Updated dependencies
+  - @pikas-ui/styles@2.3.2
+  - @pikas-ui/tooltip@2.1.2
+  - @pikas-ui/icons@2.1.2
+  - @pikas-ui/loader@2.1.2
+  - @pikas-ui/text@3.1.2
+
 ## 2.1.2
 
 ### Patch Changes